LPS-109946 Move getBy to LiferaySeleniumUtil and consolidate usages
