+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/selenium/BaseWebDriverImpl.java
@@ -3458,42 +3458,7 @@ public abstract class BaseWebDriverImpl implements LiferaySelenium, WebDriver {
 	}
 
 	protected By getBy(String locator) {
-		if (locator.startsWith("//")) {
-			return By.xpath(locator);
-		}
-		else if (locator.startsWith("class=")) {
-			locator = locator.substring(6);
-
-			return By.className(locator);
-		}
-		else if (locator.startsWith("css=")) {
-			locator = locator.substring(4);
-
-			return By.cssSelector(locator);
-		}
-		else if (locator.startsWith("link=")) {
-			locator = locator.substring(5);
-
-			return By.linkText(locator);
-		}
-		else if (locator.startsWith("name=")) {
-			locator = locator.substring(5);
-
-			return By.name(locator);
-		}
-		else if (locator.startsWith("tag=")) {
-			locator = locator.substring(4);
-
-			return By.tagName(locator);
-		}
-		else if (locator.startsWith("xpath=") || locator.startsWith("xPath=")) {
-			locator = locator.substring(6);
-
-			return By.xpath(locator);
-		}
-		else {
-			return By.id(locator);
-		}
+		return LiferaySeleniumUtil.getBy(locator);
 	}
 
 	protected Condition getConfirmationCondition(String pattern) {
